rev combat phase 4.x
开发相关：
- 为战斗流程添加了说明文档（大概吧）；
- 没有使用命名空间；
