+++ gamedata/cache/gamecfg_1.php
@@ -101,7 +101,7 @@ $trap_min_obbs = 1;
 //强制踩陷阱最大几率(百分比)
 $trap_max_obbs = 40;
 //道具发现基础几率(百分比);
-$item_obbs = 60;
+$item_obbs = 70;
 //敌人发现基础几率(百分比)
 $enemy_obbs = 70;
 //尸体发现几率（百分比）
