+++ include/game.func.php
@@ -120,8 +120,8 @@ function init_profile(){
 	}
 
 	include_once GAME_ROOT.'./include/game/revattr.func.php';
-	$atkinfo = get_base_att($pdata,$pdata,1,1);
-	$definfo = get_base_def($pdata,$pdata,1,1);
+	$atkinfo = \revattr\get_base_att($pdata,$pdata,1,1);
+	$definfo = \revattr\get_base_def($pdata,$pdata,1,1);
 
 	return;
 }
@@ -645,10 +645,15 @@ function fetch_playerdata_by_name($n)
 	$result = $db->query("SELECT * FROM {$tablepre}players WHERE name = '$n' AND type = 0");
 	if(!$db->num_rows($result)) return NULL;
 	$data = $db->fetch_array($result);
+
+	# 转码clbpara
 	if(!empty($data['clbpara'])) $data['clbpara'] = get_clbpara($data['clbpara']);
-	//套装效果刷新
+	# 判断时效性状态
+	if(!empty($data['clbpara']['lasttimes'])) check_skilllasttimes($data);
+	# 读取套装效果
 	include_once GAME_ROOT.'./include/game/itemmain.func.php';
 	reload_set_items($pdata);
+
 	return $data;
 }
 //通过pid抓取指定玩家/NPC数据
@@ -658,10 +663,15 @@ function fetch_playerdata_by_pid($pid)
 	$result = $db->query("SELECT * FROM {$tablepre}players WHERE pid = '$pid'");
 	if(!$db->num_rows($result)) return NULL;
 	$data = $db->fetch_array($result);
+
+	# 转码clbpara
 	if(!empty($data['clbpara'])) $data['clbpara'] = get_clbpara($data['clbpara']);
-	//套装效果刷新
+	# 判断时效性状态
+	if(!empty($data['clbpara']['lasttimes'])) check_skilllasttimes($data);
+	# 读取套装效果
 	include_once GAME_ROOT.'./include/game/itemmain.func.php';
-	reload_set_items($data);
+	reload_set_items($pdata);
+
 	return $data;
 }
 //用于将指定player数据存回数据库
@@ -689,7 +699,7 @@ function get_safe_plslist($mode=1)
 	return $r;
 }
 
-function w_save($id){
+/*function w_save($id){
 	global $db,$tablepre,$w_name,$w_pass,$w_type,$w_endtime,$w_deathtime,$w_gd,$w_sNo,$w_icon,$w_club,$w_hp,$w_mhp,$w_sp,$w_msp,$w_att,$w_def,$w_pls,$w_lvl,$w_exp,$w_money,$w_bid,$w_inf,$w_rage,$w_pose,$w_tactic,$w_killnum,$w_state,$w_wp,$w_wk,$w_wg,$w_wc,$w_wd,$w_wf,$w_teamID,$w_teamPass,$w_wep,$w_wepk,$w_wepe,$w_weps,$w_arb,$w_arbk,$w_arbe,$w_arbs,$w_arh,$w_arhk,$w_arhe,$w_arhs,$w_ara,$w_arak,$w_arae,$w_aras,$w_arf,$w_arfk,$w_arfe,$w_arfs,$w_art,$w_artk,$w_arte,$w_arts,$w_itm0,$w_itmk0,$w_itme0,$w_itms0,$w_itm1,$w_itmk1,$w_itme1,$w_itms1,$w_itm2,$w_itmk2,$w_itme2,$w_itms2,$w_itm3,$w_itmk3,$w_itme3,$w_itms3,$w_itm4,$w_itmk4,$w_itme4,$w_itms4,$w_itm5,$w_itmk5,$w_itme5,$w_itms5,$w_itm6,$w_itmk6,$w_itme6,$w_itms6,$w_wepsk,$w_arbsk,$w_arhsk,$w_arask,$w_arfsk,$w_artsk,$w_itmsk0,$w_itmsk1,$w_itmsk2,$w_itmsk3,$w_itmsk4,$w_itmsk5,$w_itmsk6,$w_rp,$w_action,$w_achievement,$w_skillpoint;
 	
 	$db->query("UPDATE {$tablepre}players SET name='$w_name',pass='$w_pass',type='$w_type',endtime='$w_endtime',deathtime='$w_deathtime',gd='$w_gd',sNo='$w_sNo',icon='$w_icon',club='$w_club',hp='$w_hp',mhp='$w_mhp',sp='$w_sp',msp='$w_msp',att='$w_att',def='$w_def',pls='$w_pls',lvl='$w_lvl',exp='$w_exp',money='$w_money',bid='$w_bid',inf='$w_inf',rage='$w_rage',pose='$w_pose',tactic='$w_tactic',state='$w_state',killnum='$w_killnum',action='$w_action',wp='$w_wp',wk='$w_wk',wg='$w_wg',wc='$w_wc',wd='$w_wd',wf='$w_wf',teamID='$w_teamID',teamPass='$w_teamPass',wep='$w_wep',wepk='$w_wepk',wepe='$w_wepe',weps='$w_weps',wepsk='$w_wepsk',arb='$w_arb',arbk='$w_arbk',arbe='$w_arbe',arbs='$w_arbs',arbsk='$w_arbsk',arh='$w_arh',arhk='$w_arhk',arhe='$w_arhe',arhs='$w_arhs',arhsk='$w_arhsk',ara='$w_ara',arak='$w_arak',arae='$w_arae',aras='$w_aras',arask='$w_arask',arf='$w_arf',arfk='$w_arfk',arfe='$w_arfe',arfs='$w_arfs',arfsk='$w_arfsk',art='$w_art',artk='$w_artk',arte='$w_arte',arts='$w_arts',artsk='$w_artsk',itm0='$w_itm0',itmk0='$w_itmk0',itme0='$w_itme0',itms0='$w_itms0',itmsk0='$w_itmsk0',itm1='$w_itm1',itmk1='$w_itmk1',itme1='$w_itme1',itms1='$w_itms1',itmsk1='$w_itmsk1',itm2='$w_itm2',itmk2='$w_itmk2',itme2='$w_itme2',itms2='$w_itms2',itmsk2='$w_itmsk2',itm3='$w_itm3',itmk3='$w_itmk3',itme3='$w_itme3',itms3='$w_itms3',itmsk3='$w_itmsk3',itm4='$w_itm4',itmk4='$w_itmk4',itme4='$w_itme4',itms4='$w_itms4',itmsk4='$w_itmsk4',itm5='$w_itm5',itmk5='$w_itmk5',itme5='$w_itme5',itms5='$w_itms5',itmsk5='$w_itmsk5',itm6='$w_itm6',itmk6='$w_itmk6',itme6='$w_itme6',itms6='$w_itms6',itmsk6='$w_itmsk6',rp='$w_rp',achievement='$w_achievement',skillpoint='$w_skillpoint'    WHERE pid='$id'");
@@ -706,6 +716,41 @@ function w_save2(&$data){
 		//$db->query("UPDATE {$tablepre}players SET name='$w_name',pass='$w_pass',type='$w_type',endtime='$w_endtime',gd='$w_gd',sNo='$w_sNo',icon='$w_icon',club='$w_club',hp='$w_hp',mhp='$w_mhp',sp='$w_sp',msp='$w_msp',att='$w_att',def='$w_def',pls='$w_pls',lvl='$w_lvl',exp='$w_exp',money='$w_money',bid='$w_bid',inf='$w_inf',rage='$w_rage',pose='$w_pose',tactic='$w_tactic',state='$w_state',killnum='$w_killnum',wp='$w_wp',wk='$w_wk',wg='$w_wg',wc='$w_wc',wd='$w_wd',wf='$w_wf',teamID='$w_teamID',teamPass='$w_teamPass',wep='$w_wep',wepk='$w_wepk',wepe='$w_wepe',weps='$w_weps',wepsk='$w_wepsk',arb='$w_arb',arbk='$w_arbk',arbe='$w_arbe',arbs='$w_arbs',arbsk='$w_arbsk',arh='$w_arh',arhk='$w_arhk',arhe='$w_arhe',arhs='$w_arhs',arhsk='$w_arhsk',ara='$w_ara',arak='$w_arak',arae='$w_arae',aras='$w_aras',arask='$w_arask',arf='$w_arf',arfk='$w_arfk',arfe='$w_arfe',arfs='$w_arfs',arfsk='$w_arfsk',art='$w_art',artk='$w_artk',arte='$w_arte',arts='$w_arts',artsk='$w_artsk',itm0='$w_itm0',itmk0='$w_itmk0',itme0='$w_itme0',itms0='$w_itms0',itmsk0='$w_itmsk0',itm1='$w_itm1',itmk1='$w_itmk1',itme1='$w_itme1',itms1='$w_itms1',itmsk1='$w_itmsk1',itm2='$w_itm2',itmk2='$w_itmk2',itme2='$w_itme2',itms2='$w_itms2',itmsk2='$w_itmsk2',itm3='$w_itm3',itmk3='$w_itmk3',itme3='$w_itme3',itms3='$w_itms3',itmsk3='$w_itmsk3',itm4='$w_itm4',itmk4='$w_itmk4',itme4='$w_itme4',itms4='$w_itms4',itmsk4='$w_itmsk4',itm5='$w_itm5',itmk5='$w_itmk5',itme5='$w_itme5',itms5='$w_itms5',itmsk5='$w_itmsk5',itm6='$w_itm6',itmk6='$w_itmk6',itme6='$w_itme6',itms6='$w_itms6',itmsk6='$w_itmsk6' WHERE pid='$w_pid'");
 	}
 	return;
+}*/
+
+function addnoise($wp_kind, $wsk, $ntime, $npls, $nid1, $nid2, $nmode) 
+{
+	//在隐藏地图内不会传出声音信息
+	global $plsinfo;
+	if(!array_key_exists($npls,$plsinfo)) return;
+	
+	if ((($wp_kind == 'G') && (strpos ( $wsk, 'S' ) === false)) || ($wp_kind == 'F')) {
+		global $noisetime, $noisepls, $noiseid, $noiseid2, $noisemode;
+		$noisetime = $ntime;
+		$noisepls = $npls;
+		$noiseid = $nid1;
+		$noiseid2 = $nid2;
+		$noisemode = $nmode;
+		save_combatinfo ();
+	} elseif (strpos ( $wsk, 'd' ) !== false){
+		global $noisetime, $noisepls, $noiseid, $noiseid2, $noisemode;
+		$noisetime = $ntime;
+		$noisepls = $npls;
+		$noiseid = $nid1;
+		$noiseid2 = $nid2;
+		$noisemode = 'D';
+		save_combatinfo ();
+	}
+	if (strlen($wp_kind)>=3){
+		global $noisetime, $noisepls, $noiseid, $noiseid2, $noisemode,$wep;
+		$noisetime = $ntime;
+		$noisepls = $npls;
+		$noiseid = $nid1;
+		$noiseid2 = $nid2;
+		$noisemode = $wp_kind;
+		save_combatinfo ();
+	}
+	return;
 }
 
 //销毁尸体
@@ -720,6 +765,187 @@ function destory_corpse(&$edata)
 		player_save($edata);
 	}
 	return;
+}
+
+# NPC自动换装
+#说实话没有完全看懂，但是能跑就行
+function npc_changewep_rev(&$pa,&$pd,$acitve)
+{
+	global $now,$log;
+	global $rangeinfo,$ex_dmg_def;
+
+	if(!$pa['type'] || $pa['club'] != 98) return;
+
+	$dice = diceroll(99);
+	
+	if($dice > 50)
+	{
+		$weplist = Array();
+		$wepklist = Array($pa['wepk']); $weplist2 = Array();
+		for($i=0;$i<=6;$i++)
+		{
+			if(!empty($pa['itms'.$i]) && !empty($pa['itme'.$i]) && strpos($pa['itmk'.$i],'W')===0)
+			{
+				$weplist[] = Array($i,$pa['itm'.$i],$pa['itmk'.$i],$pa['itme'.$i],$pa['itms'.$i],$pa['itmsk'.$i]);
+				$wepklist[] = $pa['itmk'.$i];
+			}
+		}
+		if(!empty($weplist))
+		{
+			$wepklist = array_unique($wepklist);
+			$temp_pd_ex_keys = array_merge(get_equip_ex_array($pd),get_wep_ex_array($pd));
+			$wepkAI = $wepskAI = true;
+			if(!empty($temp_pd_ex_keys))
+			{
+				if(count($wepklist)<=1) $wepkAI = false;
+				foreach($temp_pd_ex_keys as $ex)
+				{
+					if(in_array($ex,Array('A','B'))) $wepkAI = false;
+					if(in_array($ex,Array('a','b'))) $wepskAI = false;
+				}
+			}
+			if($wepkAI)
+			{
+				$wepk_temp = $pa['wepk'];
+				foreach($weplist as $val)
+				{
+					if($rangeinfo[substr($val[2],1,1)] >= $rangeinfo[substr($wepk_temp,1,1)] && !in_array(substr($val[2],1,1),$temp_pd_ex_keys))
+					{
+						$weplist2[] = $val;
+					}
+				}
+				if($weplist2)
+				{
+					$weplist = $weplist2;
+				}
+			}
+			if($wepskAI && $weplist)
+			{
+				$minus = array();
+				foreach($weplist as $val)
+				{
+					foreach($ex_dmg_def as $key => $val2){
+						if(strpos($val[5],$key)!==false && !in_array($val2,$temp_pd_ex_keys)){
+							$minus[] = $val;
+						}
+					}
+				}
+				if(count($minus) < count($weplist)){
+					$weplist = array_diff($weplist,$minus);
+				}				
+			}
+		}
+		else 
+		{
+			//没有获取到可换装列表，直接返回
+			return;
+		}
+		
+		if(!empty($weplist))
+		{
+			$oldwep = $pa['wep'];
+			shuffle($weplist);
+			$chosen = $weplist[0];$c = $chosen[0];
+			# 刷新套装效果
+			reload_single_set_item($pa,'wep',$oldwep);
+			reload_single_set_item($pa,'wep',$chosen[1],1);
+			$pa['itm'.$c] = $pa['wep']; $pa['itmk'.$c] = $pa['wepk']; $pa['itmsk'.$c] = $pa['wepsk'];
+			$pa['itme'.$c] = $pa['wepe']; $pa['itms'.$c] = $pa['weps'];
+			$pa['wep'] = $chosen[1]; $pa['wepk'] = $chosen[2]; $pa['wepe'] = $chosen[3]; $pa['weps'] = $chosen[4]; $pa['wepsk'] = $chosen[5];
+			# 重载基础攻击信息
+			\revattr\get_wep_kind($pa);
+			$pa['wep_range'] = \revattr\get_wep_range($pa);
+			$pa['wep_skill'] = \revattr\get_wep_skill($pa);
+			$pa['wep_name'] = $pa['wep'];
+			$pa['change_wep_log'] = "<span class=\"yellow\">{$pa['nm']}</span>将手中的<span class=\"yellow\">{$oldwep}</span>卸下，装备了<span class=\"yellow\">{$pa['wep']}</span>！<br>";
+		}
+	}
+	return;
+}
+
+# NPC喊话
+# pa指npc pd指另一视角
+function npc_chat_rev(&$pa,&$pd,$mode='') 
+{
+	global $npcchat;
+	if(!empty($npcchat[$pa['type']][$pa['name']])) 
+	{
+		$nchat = $npcchat[$pa['type']][$pa['name']];
+		$chatcolor = $nchat['color'];
+		$npcwords = !empty($chatcolor) ? "<span class = \"{$chatcolor}\">" : '<span>';
+		switch ($mode) 
+		{
+			case 'attack' :
+				if (empty($pa['itmsk0'])) 
+				{
+					$npcwords .= "{$nchat[0]}";
+					$pa['itmsk0'] = 1;
+				}
+				elseif ($pa['hp'] > ($pa['mhp'] / 2)) 
+				{
+					$dice = rand ( 1, 2 );
+					$npcwords .= "{$nchat[$dice]}";
+				} 
+				else 
+				{
+					$dice = rand ( 3, 4 );
+					$npcwords .= "{$nchat[$dice]}";
+				}
+				break;
+			case 'defend' :
+				if (empty($pa['itmsk0']))
+				{
+					$npcwords .= "{$nchat[0]}";
+					$pa['itmsk0'] = 1;
+				}
+				elseif($pa['hp'] > ($pa['mhp'] / 2)) 
+				{
+					$dice = rand ( 5, 6 );
+					$npcwords .= "{$nchat[$dice]}";
+				} 
+				else 
+				{
+					$dice = rand ( 7, 8 );
+					$npcwords .= "{$nchat[$dice]}";
+				}
+				break;
+			case 'death' :
+				$npcwords .= "{$nchat[9]}";
+				break;
+			case 'escape' :
+				$npcwords .= "{$nchat[10]}";
+				break;
+			case 'cannot' :
+				$npcwords .= "{$nchat[11]}";
+				break;
+			case 'critical' :
+				$npcwords .= "{$nchat[12]}";
+				break;
+			case 'kill' :
+				$npcwords .= "{$pa['nm']}对{$pd['nm']}说道：{$nchat[13]}";
+				break;
+		}
+		$npcwords .= '</span><br>';
+		return $npcwords;
+	} 
+	elseif ($mode == 'death') 
+	{
+		global $lwinfo;
+		if (is_array($lwinfo[$pa['type']])) 
+		{
+			$lastword = $lwinfo[$pa['type']][$pa['name']];
+		} 
+		else 
+		{
+			$lastword = $lwinfo[$pa['type']];
+		}
+		$npcwords = "<span class=\"yellow\">“{$lastword}”</span><br>";
+		return $npcwords;
+	}
+	else 
+	{
+		return;
+	}
 }
