+++ include/game/revattr_extra.func.php
@@ -1,558 +1,17 @@
 <?php
+namespace revattr
+{
 
 	if(!defined('IN_GAME')) {
 		exit('Access Denied');
 	}
 
-	//include_once GAME_ROOT.'./include/game/dice.func.php';
 	include_once GAME_ROOT.'./include/game/itemmain.func.php';
-	//include_once GAME_ROOT.'./include/game/revclubskills.func.php';
-
-	//revattr_extra.func.php: 记录NPC特殊战斗机制...玩家战斗技也先放这里了 嘻嘻
+	
+	//revattr_extra.func.php: 记录NPC特殊战斗机制...
 	//Q：为什么要把每个NPC的特殊战斗机制都新建一个函数保存？
 	//A：也不是每个都要这么干……这个做法主要用于存在大段log、多段判定的机制，分离出来一是方便定位这个NPC的相关机制在哪个阶段执行，二是确保原流程的可读性；
 
-	# 技能判定（主动型）
-	function attr_extra_active_skills(&$pa,&$pd,$active,$sk='')
-	{
-		global $log,$cskills,$now;
-		# 检查主动技合法性
-		if(isset($pa['bskill']))
-		{
-			if(!check_skill_unlock($pa['bskill'],$pa) && !check_skill_cost($pa['bskill'],$pa))
-			{
-				$bsk = $pa['bskill'];
-				$bsk_name = $cskills[$bsk]['name'];
-				# 扣除怒气
-				$bsk_cost = get_skillvars($bsk,'ragecost');
-				if($bsk_cost) $pa['rage'] -= $bsk_cost;
-				# 成功释放主动技，应用标记
-				$pa['bskill_'.$bsk] = 1;
-				$pa['bskilllog'] = "<span class=\"lime\">{$pa['nm']}对{$pd['nm']}发动了技能「{$bsk_name}」！</span><br>";
-				# 限次技每次使用时次数+1
-				if(get_skilltags($bsk,'limit'))
-				{
-					set_skillpara($bsk,'active_t',get_skillpara($bsk,'active_t',$pa['clbpara'])+1,$pa['clbpara']);
-				}
-				# 重新判定双系武器的攻击方式
-				if(isset($cskills[$bsk]['wepk']) && !in_array($pa['wep_kind'],$cskills[$bsk]['wepk']))
-				{
-					get_wep_kind($pa,$cskills[$bsk]['wepk'][0]);
-				}
-				# 检查是否需要addnews
-				addnews($now,'bsk_'.$bsk,$pa['name'],$pd['name']);
-				# 检查是否需要进行logsave
-				if(!$pd['type'] && $pd['nm']!='你') $pd['logsave'] .= "<span class=\"yellow\">{$pa['name']}</span>对你发动了技能<span class=\"red\">「{$bsk_name}」</span>！";
-				elseif(!$pa['type'] && $pa['nm']!='你') $pa['logsave'] .= "你对<span class=\"yellow\">{$pd['name']}</span>发动了技能<span class=\"red\">「{$bsk_name}」</span>！";
-			}
-			else 
-			{
-				# 主动技不满足使用条件或来源非法，直接注销标记
-				unset($pa['bskill']);
-			}
-		}
-
-		# NPC释放战斗技 暂时先重复写一遍，之后合并到一个函数里
-		if($pa['type'] && !empty($pa['clbpara']['skill']))
-		{
-			# 打乱并遍历pa技能队列 寻找可用战斗技
-			$npc_skill = $pa['clbpara']['skill'];
-			shuffle($npc_skill);
-			foreach($npc_skill as $sk)
-			{
-				if(get_skilltags($sk,'battle') && !check_skill_unlock($sk,$pa) && !check_skill_cost($sk,$pa))
-				{
-					$bsk = $sk;
-					$bsk_name = $cskills[$bsk]['name'];
-					# 扣除怒气
-					$bsk_cost = get_skillvars($bsk,'ragecost');
-					if($bsk_cost) $pa['rage'] -= $bsk_cost;
-					# 成功释放主动技，应用标记
-					$pa['bskill_'.$bsk] = 1;
-					$pa['bskilllog'] = "<span class=\"lime\">{$pa['nm']}对{$pd['nm']}发动了技能「{$bsk_name}」！</span><br>";
-					# 限次技每次使用时次数+1
-					if(get_skilltags($bsk,'limit'))
-					{
-						set_skillpara($bsk,'active_t',get_skillpara($bsk,'active_t',$pa['clbpara'])+1,$pa['clbpara']);
-					}
-					# 检查是否需要addnews
-					addnews($now,'bsk_'.$bsk,$pa['name'],$pd['name']);
-					# 别让NPC放两个技能！
-					break;
-				}
-			}
-		}
-		return;
-	}
-
-	# 技能判定（被动型）：这里通常只判定技能是否生效，如生效，提供一个标记
-	function attr_extra_passive_skills(&$pa,&$pd,$active,$sk='')
-	{
-		global $cskills,$log;
-		if(!empty($pa['clbpara']['skill']))
-		{
-			$pa['skilllog'] = '';
-			# 遍历pa技能队列 检查是否解锁
-			foreach($pa['clbpara']['skill'] as $sk)
-			{
-				# passive、buff、inf标签技能通用判定
-				# 对于解锁技能，如果有特殊触发条件，在这里加入判定，否则会默认给一个触发标记
-				if((get_skilltags($sk,'passive') || get_skilltags($sk,'buff') || get_skilltags($sk,'inf')) && !check_skill_unlock($sk,$pa))
-				{
-					# 「猛击」特殊判定
-					if($sk == 'c1_crit')
-					{
-						$sk_dice = diceroll(99);
-						# 「偷袭」或「闷棍」技能生效时，「猛击」必定触发；
-						$sk_obbs = isset($pa['bskill_c1_stalk'])||isset($pa['bskill_c1_bjack']) ? 100 : get_skillvars('c1_crit','rate');
-						# 成功触发时
-						if($sk_dice < $sk_obbs) $pa['skill_c1_crit'] = 1;
-					}
-					# 「枭眼」特殊判定：射程不小于对方时激活效果
-					elseif($sk == 'c3_hawkeye' && $pa['wep_range'] >= $pd['wep_range'])
-					{
-						$pa['skill_c3_hawkeye'] = 1;
-					}
-					# 「护盾」特殊判定：生命值不小于指定百分比时激活一个护盾
-					elseif($sk == 'c7_shield')
-					{
-						$sk_lvl = get_skilllvl('c7_shield',$pa);
-						$hpalert = get_skillvars('c7_shield','hpalert',$sk_lvl);
-						if($pa['hp'] <= $pa['mhp']*($hpalert/100))
-						{
-							getclubskill('buff_shield',$pa['clbpara']);
-							set_skillpara('buff_shield','svar',get_skillvars('c7_shield','svar',$sk_lvl),$data['clbpara']);
-							$pa['skill_buff_shield'] = 1;
-							//$pa['skilllog'] .= "<span class='lime'>感知到危险，闪烁着淡蓝幽光的护盾自动出现在{$pa['nm']}身旁！<br></span>";
-							$log .= "<span class='lime'>感知到危险，闪烁着淡蓝幽光的护盾自动出现在{$pa['nm']}身旁！<br></span>";
-						}
-					}
-					# 「暗杀」特殊判定：输出一段破隐log，并丢失buff技能
-					elseif($sk == 'buff_assassin')
-					{
-						$pa['skill_buff_assassin'] = 1;
-						$pa['skilllog'] .= "<span class='yellow'>{$pa['nm']}从阴影中现出身形，打了{$pd['nm']}一个措手不及！</span><br>";
-						lostclubskill('buff_assassin',$pa['clbpara']);
-					}
-					# 「洞察」特殊判定：熟练度高于对方时触发
-					elseif($sk == 'c10_insight' && $pa['wep_skill'] > $pd['wep_skill'])
-					{
-						$pa['skill_c10_insight'] = 1;
-						$pa['skilllog'] .= "<span class='yellow'>{$pa['nm']}凭借丰富的经验看穿了{$pd['nm']}的破绽！</span><br>";
-					}
-					# 「海虎」特殊判定：
-					elseif($sk == 'c12_swell')
-					{
-						$sk_var = get_skillvars('c12_swell','swellr') * calc_enmity_losshpr($pa,$pd);
-						$dice = diceroll(99);
-						if($dice < $sk_var)
-						{
-							$pa['skill_c12_swell'] = $pa['hp'] <= $pa['mhp']*0.3 ? 2 : 1;
-						}
-					}
-					# 其他非特判技能，默认给一个触发标记
-					else 
-					{
-						$pa['skill_'.$sk] = 1;
-						//$pa['skill_'.$sk.'_log'] = "";
-					}
-				}
-				# switch标签技能通用判定
-				if(get_skilltags($sk,'switch') && !empty(get_skillpara($sk,'active',$pa['clbpara'])) && !check_skill_unlock($sk,$pa))
-				{
-					$pa['askill_'.$sk] = 1;
-				}
-			}
-		}
-		return;
-	}
-
-	# 获取社团技能对躲避率（pd是否躲避pa）的定值修正（新）
-	function get_clbskill_hide_rate_fix(&$pa,&$pd,$r)
-	{
-		# pd持有「潜行」时的效果判定：
-		if(!check_skill_unlock('c5_sneak',$pd))
-		{
-			$sk_lvl = get_skilllvl('c5_sneak',$pd);
-			$sk_r = get_skillvars('c5_sneak','hidegain',$sk_lvl);
-			$r += $sk_r;
-		}
-		# pd持有「天助」时的效果判定：
-		if(!check_skill_unlock('c6_godsend',$pd))
-		{
-			$sk_r = get_skilllvl('c6_godsend','hidegain',$pd['clbpara']);
-			if(!empty($sk_r)) $r += $sk_r;
-		}
-		# pd处于「暗杀」状态下的效果判定：
-		if(!check_skill_unlock('buff_assassin',$pd))
-		{
-			$sk_r = get_skillvars('buff_assassin','hidegain');
-			$r += $sk_r;
-		}
-		# pa持有「瞩目」时的效果判定：
-		if(!check_skill_unlock('c12_huge',$pa))
-		{
-			$sk_r = get_skillvars('c12_huge','hideloss');
-			$r -= $sk_r;
-		}
-		# pd持有「瞩目」时的效果判定：
-		if(!check_skill_unlock('c12_huge',$pd))
-		{
-			$sk_r = get_skillvars('c12_huge','hidegain');
-			$r -= $sk_r;
-		}
-		return $r;
-	}
-
-	# 获取社团技能对先攻率（pa是否先攻pd）的定值修正（新）
-	function get_clbskill_active_rate_fix(&$pa,&$pd,$r)
-	{
-		# pa持有「枭眼」时的效果判定：
-		if(!check_skill_unlock('c3_hawkeye',$pa))
-		{
-			//计算双方射程差
-			$pa['wep_range'] = get_wep_range($pa); 
-			$pd['wep_range'] = get_wep_range($pd); 
-			if($pa['wep_range'] >= $pd['wep_range'])
-			{
-				$sk_r = get_skillvars('c3_hawkeye','actgain');
-				$r += $sk_r;
-			}
-		}
-		# pd持有「枭眼」时的效果判定：
-		if(!check_skill_unlock('c3_hawkeye',$pd))
-		{
-			//计算双方射程差
-			$pa['wep_range'] = get_wep_range($pa); 
-			$pd['wep_range'] = get_wep_range($pd); 
-			if($pa['wep_range'] < $pd['wep_range'])
-			{
-				$sk_r = get_skillvars('c3_hawkeye','actgain');
-				$r -= $sk_r;
-			}
-		}
-		# pa持有「潜行」时的效果判定：（只在主动发现敌人时应用）
-		if(!check_skill_unlock('c5_sneak',$pa))
-		{
-			$sk_lvl = get_skilllvl('c5_sneak',$pa);
-			$sk_r = get_skillvars('c5_sneak','actgain',$sk_lvl);
-			$r += $sk_r;
-		}
-		# pa持有「天助」时的效果判定：（只在主动发现敌人时应用）
-		if(!check_skill_unlock('c6_godsend',$pa))
-		{
-			$sk_r = get_skillpara('c6_godsend','actgain',$pa['clbpara']);
-			if(!empty($sk_r)) $r += $sk_r;
-		}
-		# pa处于「暗杀」状态下的效果判定：
-		if(!check_skill_unlock('buff_assassin',$pa))
-		{
-			$sk_r = get_skillvars('buff_assassin','actgain');
-			$r += $sk_r;
-		}
-		# pa持有「洞察」时的效果判定：（只在主动发现敌人时应用）
-		if(!check_skill_unlock('c10_insight',$pa) && get_wep_skill($pa) > get_wep_skill($pd))
-		{
-			$sk_lvl = get_skilllvl('c10_insight',$pa);
-			$sk_r = get_skillvars('c10_insight','actgain',$sk_lvl);
-			$r += $sk_r;
-		}
-		return $r;
-	}
-
-	# 获取社团技能对基础反击率的修正（新）
-	function get_clbskill_counterate(&$pa,&$pd,$active,$counterate)
-	{
-		#「直感」效果判定：
-		if(isset($pa['skill_c2_intuit']))
-		{
-			$sk_lvl = get_skilllvl('c2_intuit',$pa);
-			//获取反击倍率加成
-			$sk_r = 1 + (get_skillvars('c2_intuit','countergain',$sk_lvl) / 100);
-			$counterate *= $sk_r;
-		}
-		#「臂力」效果判定：
-		if(isset($pa['skill_c3_pitchpow']))
-		{
-			$sk_lvl = get_skilllvl('c3_pitchpow',$pa);
-			//获取反击倍率加成
-			$sk_r = 1 + (get_skillvars('c3_pitchpow','countergain',$sk_lvl) / 100);
-			$counterate *= $sk_r;
-		}
-		#「天助」效果判定：
-		if(isset($pa['skill_c6_godsend']))
-		{
-			$sk_r = get_skillpara('c6_godsend','countergain',$pa['clbpara']);
-			if(!empty($sk_r))
-			{
-				$sk_r = 1 + ($sk_r / 100);
-				$counterate *= $sk_r;
-			}
-		}
-		return $counterate;
-	}
-
-	# 获取社团技能对基础命中率的修正（新）
-	function get_clbskill_hitrate(&$pa,&$pd,$active,$hitrate)
-	{
-		# 加成：
-		#「潜能」、「暗杀」必中效果判定：
-		if(isset($pa['bskill_c3_potential']) || isset($pa['skill_buff_assassin']))
-		{
-			//原来必中是这个意思……
-			return 10000;
-		}
-		#「直感」效果判定：
-		if(isset($pa['skill_c2_intuit']))
-		{
-			$sk_lvl = get_skilllvl('c2_intuit',$pa);
-			$sk_r = 1 + (get_skillvars('c2_intuit','accgain',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「静息」效果判定：
-		if(isset($pa['skill_c4_stable']))
-		{
-			$sk_lvl = get_skilllvl('c4_stable',$pa);
-			$sk_r = 1 + (get_skillvars('c4_stable','accgain',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「瞄准」效果判定：
-		if(isset($pa['bskill_c4_aiming']))
-		{
-			$sk_r = 1 + (get_skillvars('c4_aiming','accgain') / 100);
-			$hitrate *= $sk_r;
-		}
-		#「穿杨」效果判定：
-		if(isset($pa['bskill_c4_sniper']))
-		{
-			$sk_r = 1 + (get_skillvars('c4_sniper','accgain') / 100);
-			$hitrate *= $sk_r;
-		}
-		#「天运」效果判定：
-		if(isset($pa['skill_c6_godluck']))
-		{
-			$sk_r = get_skillpara('c6_godluck','accgain',$pa['clbpara']);
-			if(!empty($sk_r))
-			{
-				$sk_r = 1 + ($sk_r / 100);
-				$hitrate *= $sk_r;
-			}
-		}
-		#「洞察」效果判定：
-		if(isset($pa['skill_c10_insight']))
-		{
-			$sk_lvl = get_skilllvl('c10_insight',$pa);
-			$sk_r = 1 + (get_skillvars('c10_insight','accgain',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-
-		# 减益：
-		#「枭眼」效果判定：
-		if(isset($pd['skill_c3_hawkeye']))
-		{
-			$sk_r = 1 - (get_skillvars('c3_hawkeye','accloss') / 100);
-			$hitrate *= $sk_r;
-		}
-		#「灵力」效果判定：
-		if(isset($pd['skill_c9_spirit']))
-		{
-			$sk_lvl = get_skilllvl('c9_spirit',$pd);
-			$sk_r = 1 - (get_skillvars('c9_spirit','accloss',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「天运」效果判定：
-		if(isset($pd['skill_c6_godluck']))
-		{
-			$sk_r = get_skillpara('c6_godluck','accloss',$pd['clbpara']);
-			if(!empty($sk_r))
-			{
-				$sk_r = 1 - ($sk_r / 100);
-				$hitrate *= $sk_r;
-			}
-		}
-		#「洞察」效果判定：
-		if(isset($pd['skill_c10_insight']))
-		{
-			$sk_lvl = get_skilllvl('c10_insight',$pd);
-			$sk_r = 1 - (get_skillvars('c10_insight','accloss',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		return $hitrate;
-	}
-
-	# 获取社团技能对连击命中率的修正（新）
-	function get_clbskill_r_hitrate(&$pa,&$pd,$active,$hitrate)
-	{
-		# 加成：
-		#「潜能」效果判定：
-		if(isset($pa['bskill_c3_potential']))
-		{
-			//潜能激活时连击无衰减
-			return 1;
-		}
-		#「直感」效果判定：
-		if(isset($pa['skill_c2_intuit']))
-		{
-			$sk_lvl = get_skilllvl('c2_intuit',$pa);
-			//获取连击命中率加成
-			$sk_r = 1 + (get_skillvars('c2_intuit','rbgain',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「静息」效果判定：
-		if(isset($pa['skill_c4_stable']))
-		{
-			$sk_lvl = get_skilllvl('c4_stable',$pa);
-			//获取连击命中率加成
-			$sk_r = 1 + (get_skillvars('c4_stable','rbgain',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「天运」效果判定：
-		if(isset($pa['skill_c6_godluck']))
-		{
-			$sk_r = get_skillpara('c6_godluck','rbgain',$pa['clbpara']);
-			if(!empty($sk_r))
-			{
-				$sk_r = 1 + ($sk_r / 100);
-				$hitrate *= $sk_r;
-			}
-		}
-
-		# 减益：
-		#「枭眼」效果判定：
-		if(isset($pd['skill_c3_hawkeye']))
-		{
-			$sk_r = 1 - (get_skillvars('c3_hawkeye','rbloss') / 100);
-			$hitrate *= $sk_r;
-		}
-		#「灵力」效果判定：
-		if(isset($pd['skill_c9_spirit']))
-		{
-			$sk_lvl = get_skilllvl('c9_spirit',$pd);
-			$sk_r = 1 - (get_skillvars('c9_spirit','rbloss',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		#「天运」效果判定：
-		if(isset($pd['skill_c6_godluck']))
-		{
-			$sk_r = get_skillpara('c6_godluck','rbloss',$pd['clbpara']);
-			if(!empty($sk_r))
-			{
-				$sk_r = 1 - ($sk_r / 100);
-				$hitrate *= $sk_r;
-			}
-		}
-		#「洞察」效果判定：
-		if(isset($pd['skill_c10_insight']))
-		{
-			$sk_lvl = get_skilllvl('c10_insight',$pd);
-			$sk_r = 1 - (get_skillvars('c10_insight','rbloss',$sk_lvl) / 100);
-			$hitrate *= $sk_r;
-		}
-		return $hitrate;
-	}
-
-	# 获取社团技能对致伤率（防具耐久损伤率）的修正（新）
-	function get_clbskill_infrate(&$pa,&$pd,$active,$infrate)
-	{
-		#「破甲」效果判定：
-		if(isset($pa['skill_c4_break']))
-		{
-			$sk_lvl = get_skilllvl('c4_break',$pa);
-			//获取致伤率加成
-			$sk_r = 1 + (get_skillvars('c4_break','infrgain',$sk_lvl) / 100);
-			$infrate *= $sk_r;
-		}
-		return $infrate;
-	}
-
-	# 获取社团技能对基础致伤效果（每次致伤会损耗多少点防具耐久）的修正（新）
-	function get_clbskill_inftimes(&$pa,&$pd,$active,$inftimes)
-	{
-		#「破甲」效果判定：
-		if(isset($pa['skill_c4_break']))
-		{
-			$sk_lvl = get_skilllvl('c4_break',$pa);
-			//获取致伤效果加成
-			$sk_fix = get_skillvars('c4_break','inftfix',$sk_lvl);
-			$inftimes += $sk_fix;
-		}
-		#「咆哮」效果判定：
-		if(isset($pa['bskill_c4_roar']))
-		{
-			//获取致伤效果加成
-			$sk_fix = get_skillvars('c4_roar','inftfix');
-			$inftimes += $sk_fix;
-		}
-		return $inftimes;
-	}
-
-	# 获取社团技能对伤害浮动的修正（新）
-	function get_clbskill_fluc(&$pa,&$pd,$active)
-	{
-		#「直感」效果判定：
-		if(isset($pa['skill_c2_intuit']))
-		{
-			$sk_lvl = get_skilllvl('c2_intuit',$pa);
-			//获取伤害浮动加成
-			$sk_fix = get_skillvars('c2_intuit','flucgain',$sk_lvl);
-			return $sk_fix;
-		}
-		return 0;
-	}
-
-	# 计算社团技能对单个属性基础伤害的系数补正
-	function get_clbskill_ex_base_dmg_r(&$pa,&$pd,$active,$key)
-	{
-		$ex_dmg_r = 1;
-		# 「附魔」效果判定：
-		if(isset($pa['bskill_c3_enchant']))
-		{
-			$exdmgarr = get_skillvars('c3_enchant','exdmgarr');
-			if(isset($exdmgarr[$key]) && !empty(get_skillpara('c3_enchant',$exdmgarr[$key],$pa['clbpara'])))
-			{
-				$ex_r = get_skillpara('c3_enchant',$exdmgarr[$key],$pa['clbpara']);
-				$ex_dmg_r += $ex_r/100;
-				//echo "【DEBUG】附魔使{$key}伤害提高了{$ex_r}%<br>";
-			}
-		}
-		return $ex_dmg_r;
-	}
-
-	# 获取社团技能对单个属性基础伤害的定值补正
-	function get_clbskill_ex_base_dmg_fix(&$pa,&$pd,$active,$key)
-	{
-		$ex_dmg_fix = 0;
-		# 「歼灭」效果判定：
-		if(isset($pa['skill_buff_annihil']))
-		{
-			global $ex_wep_dmg;
-			$sk_var = round($pa['att']/$ex_wep_dmg[$key]);
-			$ex_dmg_fix += $sk_var;
-		}
-		# 「磁暴」效果判定：
-		if(isset($pa['bskill_c7_electric']) && $key == 'e')
-		{
-			$sk_var = get_skillvars('c7_electric','exdmgfix');
-			$ex_damage_fix += $sk_var;
-		}
-		return $ex_dmg_fix;
-	}
-
-	# 背水HP系数
-	function calc_enmity_losshpr(&$pa,&$pd)
-	{
-		$hpr = 1 - ($pa['hp']/$pa['mhp']);
-		$r = (1 + 2*$hpr) * $hpr;
-		return $r;
-	}
-
-	# 坚守HP系数
-	function calc_garrison_losshpr(&$pa,&$pd)
-	{
-		$hpr = 1 - ($pa['hp']/$pa['mhp']);
-		$r = -1 * pow($hpr,3) + 4 * $hpr;
-		return $r;
-	}
 
 	# 真红暮特殊判定
 	function attr_extra_19_crimson(&$pa,&$pd,$active,$phase=0)
@@ -1013,5 +472,5 @@
 		}
 		return;
 	}
-
+}
 ?>
